refactor: resolve circular imports by inlining file adapter
 - while it's perfectly fine to use lazy imports to workaround circular references, it appears to destroy the pylance integration and auto-imports.
 - move the clients.py and file.py code into base to make it so there is no need for lazy import workarounds.
