+++ pathy/__init__.py
@@ -1,25 +1,32 @@
 from .base import (
-    BasePath,
-    Blob,
-    BlobStat,
-    Bucket,
-    BucketClient,
-    BucketEntry,
-    BucketsAccessor,
-    ClientError,
-    FluidPath,
-    Pathy,
-    PurePathy,
+    BasePath as BasePath,
+    Blob as Blob,
+    BlobStat as BlobStat,
+    Bucket as Bucket,
+    BucketClient as BucketClient,
+    BucketEntry as BucketEntry,
+    BucketsAccessor as BucketsAccessor,
+    ClientError as ClientError,
+    FluidPath as FluidPath,
+    Pathy as Pathy,
+    PurePathy as PurePathy,
+    BlobFS as BlobFS,
+    BucketClientFS as BucketClientFS,
+    BucketEntryFS as BucketEntryFS,
+    BucketFS as BucketFS,
+    clear_fs_cache as clear_fs_cache,
+    get_client as get_client,
+    get_fs_cache as get_fs_cache,
+    get_fs_client as get_fs_client,
+    register_client as register_client,
+    set_client_params as set_client_params,
+    use_fs as use_fs,
+    use_fs_cache as use_fs_cache,
 )
-from .clients import (
-    clear_fs_cache,
-    get_client,
-    get_fs_cache,
-    get_fs_client,
-    register_client,
-    set_client_params,
-    use_fs,
-    use_fs_cache,
+from .gcs import (
+    BlobGCS as BlobGCS,
+    BucketClientGCS as BucketClientGCS,
+    BucketEntryGCS as BucketEntryGCS,
+    BucketGCS as BucketGCS,
+    has_gcs as has_gcs,
 )
-from .file import BlobFS, BucketClientFS, BucketEntryFS, BucketFS
-from .gcs import BlobGCS, BucketClientGCS, BucketEntryGCS, BucketGCS, has_gcs
